Fix fatal bug in Win32 constructor
During the recent changes to Win32 constructor, I intended to
have the 'command' added to the options in all cases, but I
never followed through with that in Background.pm.  On further
thought, this would require cloning the options hash so might
as well just leave it how it is with the command and exe as
attributes of the object.